Create example function to be shown in godoc
Example functions are treated like tests and the godoc tool can
automatically extract them to show on documentation.

It is important to note that the comment inside the example function is
of extreme importance. Without it, the code will be compiled but not
executed as a test. Also, it will be shown in documentation, but the
expected output will be missing.