Make function that return sum of values in slices
Lot of new things here. First, there is the ...[]int in the function
arguments. This is what is called a variadic function. It allows your
function to take any number of arguments of specified type. In this
case, my function accept multiple slices. If I had a function like
func sum(nums ...int) and my integers were inside a slice numbers{1,2,3}, I
could pass it like this: sum(numbers...)

There is also the reflect.DeepEqual function. This allows us to be a
little more flexible with comparisons, to be used where operator ==
would be a hassle to implement. It can be inconsistent, so use with
caution

The range function here will be iterating over the numbersToSum, so each
element "numbers" will be a slice of integers, that we already know how
to sum with the function Sum. I'm not sure about the type of
lenghtOfNumbers as of now, but will investigate it later.