Add struct Rectangle
This make our code smaller, and clearer since we don't need to name our
functions 'XxxArea' if we want to calculate area and perimeters of other
functions. This way we can create new structures for new shapes if we
want, and just modify the functions to deal with these new shapes.

There is a better way to do this, though: by using methods. This is what
I will study next.