Refactor using methods to make adding shapes easy
Here we use structs and methods so each shape have it's own methods to
deal with calculating the Area and Perimeter. This way, we make adding
new shapes easy, and each shape has it's own methods so we don't have
to make a function deal with different structs and formulas to calculate
each area/perimeter.