Make a basic wallet with structs and pointers
Creating a basic struct to represent a wallet that will recieve bitcoins

We use pointers when passing the wallet to the methods of this struct,
since the arguments are copied when passed to a function and that is not
what I want here, I want a single structures with fields I can modify
and access via methods. Passing a pointer to the structure make the code
behave the way I want it.

Note that I didn't dereference the parameter with (*w).balance as
expected. That's because w is a struct pointer, which have automatic
dereference in Go. It is like that just to make notation better, but I
am still accessing the value the pointer points to.